Update README.md
Remove racism. "golliwog's cakewalk song" music example does not recognize that a "golliwog" doll is a racialized caricature. Debussy was born well past the start of the trans-Atlantic slave trade. The term was used as a slur before his death, and so the title choice was either not self-aware or in denial of what the word actually represents.